[add] 添加了基于catch2的单元测试项目

+++ project/virtual-disk-system/node/FileNode.h
@@ -12,7 +12,7 @@ public:
 
 	string_local GetDataByCopy();
 	void SetDataByCopy(const string_local str);
-	virtual uint64_t GetSize();
+	virtual size_t GetSize();
 
 private:
 	string_local m_file_data;
