+++ project/virtual-disk-system/node/SymlinkNode.h
@@ -10,7 +10,7 @@ public:
 	explicit SymlinkNode(string_local name, BaseNode* parent = nullptr);
 	virtual ~SymlinkNode();
 	void SetSymlink(NodeType type, BaseNode* dst_node);	//设置链接对象
-	virtual uint64_t GetSize();
+	virtual size_t GetSize();
 private:
 	BaseNode* m_dst_node = nullptr;
 };
